+++ NEWS.md
@@ -2,6 +2,8 @@
 
 ### Changed
 
+* rename variable `t` with `t_anom`
+* add variable `t_corr` discard correlation threshold
 * define `discard_anomalous_rules()` and `discard_corre_rules()` functions
 * reorganize `generate_rules_matrix()` (separate standardization, and remove filtering)
 * explicit `prune_rules()` function and add relative tests